use DropdownMenu instead of CategoryDropdown
