+++ app/components/search/Header/Header.tsx
@@ -5,7 +5,7 @@ import { Button } from "components/ui/inline/Button/Button";
 import { QrCodeModal } from "components/ui/QrCodeModal/QrCodeModal"; // todo: remove QrCodeModal from project
 import websiteConfig from "utils/websiteConfig";
 import { CATEGORIES } from "pages/constants";
-import { CategoryDropdown } from "../../ui/Navigation/CategoryDropdown";
+import DropdownMenu from "components/ui/Navigation/DropdownMenu";
 
 import styles from "./Header.module.scss";
 
@@ -22,6 +22,19 @@ export const Header = ({
 
   const title = resultsTitle === "" ? "All categories" : resultsTitle;
 
+  const links = [
+    {
+      id: "all-categories",
+      url: "/search",
+      text: "All categories",
+    },
+    ...CATEGORIES.map((category) => ({
+      id: category.slug,
+      url: `/${category.slug}/results`,
+      text: category.name,
+    })),
+  ];
+
   return (
     <div className={styles.header}>
       <div className={styles.headerInner}>
@@ -29,7 +42,12 @@ export const Header = ({
           <h1 className="sr-only">
             {title === resultsTitle ?? "Search results"}
           </h1>
-          <CategoryDropdown categories={CATEGORIES} resultsTitle={title} />
+          <DropdownMenu
+            title={resultsTitle}
+            links={links}
+            uniqueKey={resultsTitle}
+            variant="category"
+          />
         </div>
         <Button
           onClick={() => {
